+++ frontend/src/frontend/config/events.cljs
@@ -9,19 +9,20 @@
    [modular.encoding.transit :refer [decode]]
    [modular.encoding.edn :refer [read-edn]]
    [frontend.notifications.core :refer [add-notification]]
-   [frontend.routes :refer [static-main-path-atom]]
+   [webly.app.mode :refer [get-resource-path get-mode]]
    ))
 
 ; load configuration
 
 (reg-event-fx
  :config/load
- (fn [{:keys [db]} [_ after-config-load static?]]
-   (let [format (if static?
+ (fn [{:keys [db]} [_ after-config-load]]
+   (let [static? (= (get-mode) :static)
+         format (if static?
                   (ajax/text-response-format)
                   (ajax/transit-response-format :json decode))
          uri (if static?
-               (str @static-main-path-atom "/r/config.edn")
+               (str (get-resource-path) "config.edn")
                "/api/config")]
      (infof "loading config static?: %s from url: %s then dispatch: %s" static? uri after-config-load)
      {;:db   (assoc-in db [:build] (pref))