+++ webly/src/webly/app/mode/url.cljs
@@ -7,12 +7,13 @@
 (defn application-url []
   (-> js/window .-location .-href))
 
-(defn dynamic-base []
+(defn app-load-path []
   (let [url (application-url)
         url-base (subs url 0 (last-index-of url "/"))]
-    (info "dynamic-base: " url-base)
+    (info "app-load-path: " url-base)
     url-base))
 
+(def dynamic-base app-load-path)
 
 (defn entry-path []
   (let [{:keys [protocol port host path]} (application-url)]