+++ CHANGELOG.md
@@ -7,3 +7,5 @@
 - ring-handlers get required on startup (app will not start with invalid handlers)
 - cljs-build uses extensions (can use all ui-extensions without sci)
 - runtime services using extensions (api-routes, cljs-routes, theme)
+- webly profile: "static": writes all resources neded for a static page
+- static build working.